refactor: move albums logic to albums_service.py
