+++ app/main.py
@@ -1,19 +1,39 @@
 import webview
 
 from app.api.auth_api import AuthApi
+from app.icloud.albums_service import AlbumsService
 
+# --- API Bridge ---
+auth_api = AuthApi()
 
-def main():
-    api = AuthApi()
+class API:
+    def __init__(self):
+        self.albums_service = None
 
-    webview.create_window(
-        "iCloud Sorter",
-        "ui/index.html",
-        js_api=api,
-    )
+    def login(self, apple_id, password):
+        result = auth_api.login(apple_id, password)
+        if result.get("success"):
+            self.albums_service = AlbumsService(auth_api.icloud)
+        return result
 
-    webview.start()
+    def verify_2fa(self, code):
+        result = auth_api.verify_2fa(code)
+        if result.get("success"):
+            self.albums_service = AlbumsService(auth_api.icloud)
+        return result
 
+    def get_albums(self):
+        if not self.albums_service:
+            return []
+        return self.albums_service.get_albums()
 
-if __name__ == "__main__":
-    main()
+# --- Create pywebview window ---
+webview.create_window(
+    "iCloud Photo Sorter",
+    "ui/index.html",
+    js_api=API(),
+    width=800,
+    height=600,
+)
+
+webview.start()
